backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
groq==0.7.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
cors==1.0.1
aiofiles==23.2.1
